Dithering/requirements.txt
flask
flask_restplus
Pillow
numpy
